+++ doc/examples/aseprite/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
-import 'package:flame/animation.dart' as flame_animation;
-import 'package:flame/components/animation_component.dart';
+import 'package:flame/sprite_animation.dart';
+import 'package:flame/components/sprite_animation_component.dart';
 import 'package:flutter/material.dart';
 
 void main() async {
@@ -17,11 +17,11 @@ class MyGame extends BaseGame {
   }
 
   void _start() async {
-    final animation = await flame_animation.Animation.fromAsepriteData(
+    final animation = await SpriteAnimation.fromAsepriteData(
       'chopper.png',
       'chopper.json',
     );
-    final animationComponent = AnimationComponent(200, 200, animation);
+    final animationComponent = SpriteAnimationComponent(200, 200, animation);
 
     animationComponent.x = (size.width / 2) - 100;
     animationComponent.y = (size.height / 2) - 100;