+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ## [next]
  - Preventing some crashed that could happen on web when some methods were called
+ - Rename Animation to SpriteAnimation
 
 ## 0.24.0
  - Outsourcing SVG support to an external package